requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
click==8.1.3
fastapi==0.82.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
passlib==1.7.4
pydantic==1.10.2
pypika-tortoise==0.1.6
pytz==2022.2.1
sniffio==1.3.0
starlette==0.19.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.3
